requirements.txt
aioredis==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cache2==0.2.2
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
flake8==7.1.2
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.0
makefun==1.15.6
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
pendulum==3.0.0
pluggy==1.5.0
pwdlib==0.2.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
